+++ test/unit/lib/cryptoUtilTests.js
@@ -92,6 +92,20 @@ describe('cryptoUtil', () => {
                 assert.strictEqual(logWarningSpy.args[0][0], 'Failed to decrypt data with id');
             });
         });
+
+        it('should handle list command errors', () => {
+            const error = 'The requested RADIUS Server (/Common/doBigIp) was not found';
+            const logWarningSpy = sinon.spy(Logger.prototype, 'warning');
+            sinon.stub(cloudUtil, 'runTmshCommand').rejects(new Error(error));
+            return assert.isRejected(
+                cryptoUtil.decryptStoredValueById('foo', '123-abc'),
+                'The requested RADIUS Server (/Common/doBigIp) was not found',
+                'should have caught error'
+            ).then(() => {
+                assert.strictEqual(logWarningSpy.thisValues[0].metadata, 'cryptoUtil.js | 123-abc');
+                assert.strictEqual(logWarningSpy.args[0][0], 'There was no value to decrypt. This can happen if there is an unexpected restart.');
+            });
+        });
     });
 
     describe('deleteEncryptedId', () => {